pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kokoro-mcp-server"
version = "0.1.0"
description = "AI アシスタントと連携し、テキストを高品質な音声に変換する MCP サーバー"
readme = "README.md"
requires-python = ">=3.10, <3.12"
license = {file = "LICENSE"}
authors = [
    {name = "kokoro-MCP-Server Contributors"}
]
dependencies = [
    "websockets>=10.4",
    "python-dotenv>=1.0.0",
    "click>=8.1.3",
    "loguru>=0.7.0",
    "pydantic>=2.0.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "mcp-python>=0.1.0",
    "pydub>=0.25.1",
    "soundfile>=0.12.1",
    "numpy>=1.24.0,<2.0.0",
    "langdetect>=1.0.9",
    "jaconv>=0.3",
    "emoji>=2.2.0",
    "aiohttp>=3.8.4",
    "requests>=2.28.2",
    "cachetools>=5.3.0",
    "uvloop>=0.17.0;sys_platform!='win32'",
    "kokoro>=0.9.2",
    "librosa>=0.10.1",
    "torch>=2.0.0,<2.2.0",
    "pyopenjtalk>=0.4.0",
    "fugashi>=1.4.0",
    "mojimoji>=0.0.13",
    "mcp[cli]>=1.5.0",
    "wasabi>=1.1.3",
    "spacy>=3.8.4",
    "thinc>=8.3.4",
    "fugashi>=1.4.0",
    "mecab-python3>=1.0.4",
    "cutlet<=0.5.0",
    "soundfile>=0.12.1",
    "numpy>=1.24.0,<2.0.0",
    "langdetect>=1.0.9",
    "jaconv>=0.3",
    "emoji>=2.2.0",

]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-xdist>=3.5.0",
    "locust>=2.24.0",
    "httpx>=0.27.0",
]
dev = [
    "black>=24.1.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
]

[project.scripts]
kokoro-mcp-server = "kokoro_mcp_server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/kokoro_mcp_server"]

[tool.hatch.build]
packages = ["src/kokoro_mcp_server"]
only-packages = true
artifacts = [
    "src/kokoro_mcp_server/**/*.py",
    "src/kokoro_mcp_server/**/*.pyi",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
]
addopts = [
    "-v",
    "--tb=short",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-branch",
]

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false 
